SLVS-1757 Move interface into its own file
