+++ src/ConnectedMode/Suppressions/RoslynSuppressionUpdater.cs
@@ -23,40 +23,9 @@ using SonarLint.VisualStudio.ConnectedMode.Helpers;
 using SonarLint.VisualStudio.Core;
 using SonarLint.VisualStudio.Infrastructure.VS;
 using SonarQube.Client;
-using SonarQube.Client.Models;
 
 namespace SonarLint.VisualStudio.ConnectedMode.Suppressions
 {
-    /// <summary>
-    /// Fetches suppressed issues from the server and updates the store
-    /// </summary>
-    internal interface IRoslynSuppressionUpdater
-    {
-        /// <summary>
-        /// Fetches all available suppressions from the server and updates the server issues store
-        /// </summary>
-        Task UpdateAllServerSuppressionsAsync();
-
-        /// <summary>
-        /// Updates the suppression status of the given issue key(s). If the issues are not found locally, they are fetched.
-        /// </summary>
-        Task UpdateSuppressedIssuesAsync(bool isResolved, string[] issueKeys, CancellationToken cancellationToken);
-
-        event EventHandler<SuppressionsEventArgs> SuppressedIssuesReloaded;
-        event EventHandler<SuppressionsEventArgs> NewIssuesSuppressed;
-        event EventHandler<SuppressionsUpdateEventArgs> NewIssuesResolved;
-    }
-
-    public class SuppressionsEventArgs(IReadOnlyList<SonarQubeIssue> suppressedIssues) : EventArgs
-    {
-        public IReadOnlyList<SonarQubeIssue> SuppressedIssues { get; } = suppressedIssues;
-    }
-
-    public class SuppressionsUpdateEventArgs(IReadOnlyList<string> suppressedIssueKeys) : EventArgs
-    {
-        public IReadOnlyList<string> SuppressedIssueKeys { get; } = suppressedIssueKeys;
-    }
-
     [Export(typeof(IRoslynSuppressionUpdater))]
     [PartCreationPolicy(CreationPolicy.Shared)]
     internal sealed class RoslynSuppressionUpdater : IRoslynSuppressionUpdater, IDisposable
